Add yarn resolution for current nw-builder version
The grunt-nw-builder package pulls in an old version that crashes due to
an old version of the temp library. This will override it with the
latest version when using yarn. This does _not_ work with npm, you'll
still get the old one that doesn't work. It can be made to work using
some kinda plugin I think, but I'm not gonna bother for now with just me
working on it.